Fixed compiling error.  to_String can take ulong long, long double, and long long as parameter.  In this case, the use of 'order' is ambiguous.  Cast it to long long.

+++ src/gui_and_draw/GuiDevice.cpp
@@ -2278,7 +2278,7 @@ void SkinOutput::Update( int contL, int order, int contR )
     m_contL = ContStr(contL);
     m_ContLOutput->value( m_contL.c_str() );
 
-    m_order = std::to_string( order );
+    m_order = std::to_string( ( long long )order );
     m_OrderOutput->value( m_order.c_str() );
 
     m_contR = ContStr(contR);
